edit comments and replied
