--- .env
@@ -1,1 +1,0 @@
-ESLINT_NO_DEV_ERRORS = true
