+++ .gitignore
@@ -23,3 +23,4 @@ yarn-debug.log*
 yarn-error.log*
 
 /.idea
+.eslintrc.json
